attacks
attack and special attack have been implimeted, theres an issue with dash attack idk how to solve it may be that you need to check if attack is pressed while in dash, or that i need to find a way to check when im in dash and if theres a collision during that time to set hp lower. Could also be the collision mode im using. work on it